Se agregó un componente nuevo BenefitsModal, para modularizar el código del componente CrudBenefits
