+++ FrontEnd/src/Components/CrudBenefits/CrudBenefits.js
@@ -10,6 +10,7 @@ import {
 import { useState, useEffect } from "react";
 import '../../App.css'
 import "bootstrap/dist/css/bootstrap.min.css";
+import { BenefitsModal } from "./BenefitsModal";
 
 const database = [
   {
@@ -24,37 +25,6 @@ const apiBenefits = `http://localhost:5000/benefits/${projectName}`
 
 export const CrudBenefits = () => {
   const [data, setData] = useState(database);
-  const [viewModal, setViewModal] = useState(false);
-  const [name, setName] = useState('');
-  const [cost, setCost] = useState(0);
-  const [warning, setWarning] = useState('');
-
-  const addToTable = () => {
-    if (name && cost) {
-      const names = [];
-      data.map((index) => {
-        names.push(index.name);
-      })
-      if (!names.includes(name)) {
-        const newData = {
-          id: data.length + 1,
-          name: name,
-          actualCost: cost,
-        };
-        setData([...data, newData]);
-        setWarning('');
-        setViewModal(false);
-        setName("");
-        setCost("");
-      } else {
-        setWarning('*That benefit already exist')
-      }
-
-    }
-    else {
-      setWarning('*Please enter all the values')
-    }
-  }
 
   const [infoReceived, setInfoReceived] = useState(false);
   useEffect(() => {
@@ -75,10 +45,7 @@ export const CrudBenefits = () => {
     <>
       <Container className="content-container">
         <br />
-        <button className="create-button" onClick={() => setViewModal(true)}>
-          {" "}
-          Create New Benefit
-        </button>
+        <BenefitsModal data={data} setData={setData} />
         <br />
         <table className="Table">
           <thead>
@@ -105,53 +72,7 @@ export const CrudBenefits = () => {
           </tbody>
         </table>
       </Container>
-      <Modal className='modal-window' isOpen={viewModal}>
-        <ModalHeader>
-          <div>
-            <h3>Insert New Benefit</h3>
-          </div>
-        </ModalHeader>
-        <ModalBody>
-          <FormGroup>
-            <label>Name:</label>
-            <input
-              className="form-control"
-              type="text"
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-            ></input>
-          </FormGroup>
 
-          <FormGroup>
-            <label>Cost:</label>
-            <input
-              className="form-control"
-              type="number"
-              value={cost}
-              onChange={(e) => setCost(e.target.value)}
-            ></input>
-          </FormGroup>
-        </ModalBody>
-        <ModalFooter>
-          <label className="warning-message">{warning}</label>
-          <button
-            className="button create-button"
-            onClick={() => {
-              addToTable();
-            }}
-          >
-            Insert
-          </button>
-          <button className="button cancel-button" onClick={() => {
-            setViewModal(false)
-            setName("");
-            setCost("");
-            setWarning('');
-          }}>
-            Cancel
-          </button>
-        </ModalFooter>
-      </Modal>
     </>
   )
